backend/requirements.txt
aioredis==1.3.1
arrow==1.2.0
asgiref==3.3.4
astroid==2.8.2
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
constantly==15.1.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==35.0.0
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.5
django-channels-presence==1.0.0
django-cors-headers==3.7.0
django-postgrespool2==2.0.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
jinja2-time==0.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.3
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.2
pandas==1.3.4
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
pylint==2.11.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-django==4.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
requests==2.25.1
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.27
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
